use absolute path to mount and no longer prefer /var/tmp since it is the same as /tmp on OpenBSD nowadays
