+++ client/erasure_openbsd.go
